qt4-x11-free: enable -accessibility and -sm
Is it possible to enable the "-sm -accessibility" in oe-core, please?
There is a meta-kde layer which requires the "-sm -accessibility", but
they are disabled in meta/recipes-qt/qt4/qt4.inc:

QT_DISTRO_FLAGS ?= "-no-accessibility -no-sm"

I checked the log of the qt4, can't find the related log for
"-no-accessibility -no-sm".

Another way is use the bbappend, but it would be great if it can be
enabled in oe-core.

This only enables for qt4-x11, doesn't enable for qt4-embedded, and
have done testing on: qemux86, qemuarm, qemumips, qemuppc.

(From OE-Core rev: 678973a9ad258b466295a666cbcd56df15a57c7c)

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>